requirements.txt
huggingface-hub==0.23.4
numpy==1.24.3
opencv-python==4.10.0.84
pandas==2.0.2
pickleshare==0.7.5
pillow==10.3.0
pytest==8.2.2
pytest-benchmark==4.0.0
PyYAML==6.0
requests==2.31.0
timm==1.0.3
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
